Combine scheduler, todo, and overview
